Move google sign in to new file
